factor datafeedstats into separate component
